+++ pkg/proxy/proxy_test.go
@@ -25,6 +25,7 @@ import (
 
 	"github.com/brancz/kube-rbac-proxy/pkg/authn"
 	"github.com/brancz/kube-rbac-proxy/pkg/authz"
+	"github.com/google/go-cmp/cmp"
 	"k8s.io/apiserver/pkg/authentication/authenticator"
 	"k8s.io/apiserver/pkg/authentication/request/bearertoken"
 	"k8s.io/apiserver/pkg/authentication/user"
@@ -81,6 +82,176 @@ func TestProxyWithOIDCSupport(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestGeneratingAuthorizerAttributes(t *testing.T) {
+	cases := []struct {
+		desc     string
+		authzCfg *authz.Config
+		req      *http.Request
+		expected []authorizer.Attributes
+	}{
+		{
+			"without resource attributes and rewrites",
+			&authz.Config{},
+			createRequest(nil, nil),
+			[]authorizer.Attributes{
+				authorizer.AttributesRecord{
+					User:            nil,
+					Verb:            "get",
+					Namespace:       "",
+					APIGroup:        "",
+					APIVersion:      "",
+					Resource:        "",
+					Subresource:     "",
+					Name:            "",
+					ResourceRequest: false,
+					Path:            "/accounts",
+				},
+			},
+		},
+		{
+			"without rewrites config",
+			&authz.Config{ResourceAttributes: &authz.ResourceAttributes{Namespace: "tenant1", APIVersion: "v1", Resource: "namespace", Subresource: "metrics"}},
+			createRequest(nil, nil),
+			[]authorizer.Attributes{
+				authorizer.AttributesRecord{
+					User:            nil,
+					Verb:            "get",
+					Namespace:       "tenant1",
+					APIGroup:        "",
+					APIVersion:      "v1",
+					Resource:        "namespace",
+					Subresource:     "metrics",
+					Name:            "",
+					ResourceRequest: true,
+				},
+			},
+		},
+		{
+			"with query param rewrites config",
+			&authz.Config{
+				Rewrites:           &authz.SubjectAccessReviewRewrites{ByQueryParameter: &authz.QueryParameterRewriteConfig{Name: "namespace"}},
+				ResourceAttributes: &authz.ResourceAttributes{Namespace: "{{ .Value }}", APIVersion: "v1", Resource: "namespace", Subresource: "metrics"},
+			},
+			createRequest(map[string]string{"namespace": "tenant1"}, nil),
+			[]authorizer.Attributes{
+				authorizer.AttributesRecord{
+					User:            nil,
+					Verb:            "get",
+					Namespace:       "tenant1",
+					APIGroup:        "",
+					APIVersion:      "v1",
+					Resource:        "namespace",
+					Subresource:     "metrics",
+					Name:            "",
+					ResourceRequest: true,
+				},
+			},
+		},
+		{
+			"with query param rewrites config but missing URL query",
+			&authz.Config{
+				Rewrites:           &authz.SubjectAccessReviewRewrites{ByQueryParameter: &authz.QueryParameterRewriteConfig{Name: "namespace"}},
+				ResourceAttributes: &authz.ResourceAttributes{Namespace: "{{ .Value }}", APIVersion: "v1", Resource: "namespace", Subresource: "metrics"},
+			},
+			createRequest(nil, nil),
+			nil,
+		},
+		{
+			"with http header rewrites config",
+			&authz.Config{
+				Rewrites:           &authz.SubjectAccessReviewRewrites{ByHTTPHeader: &authz.HTTPHeaderRewriteConfig{Name: "namespace"}},
+				ResourceAttributes: &authz.ResourceAttributes{Namespace: "{{ .Value }}", APIVersion: "v1", Resource: "namespace", Subresource: "metrics"},
+			},
+			createRequest(nil, map[string]string{"namespace": "tenant1"}),
+			[]authorizer.Attributes{
+				authorizer.AttributesRecord{
+					User:            nil,
+					Verb:            "get",
+					Namespace:       "tenant1",
+					APIGroup:        "",
+					APIVersion:      "v1",
+					Resource:        "namespace",
+					Subresource:     "metrics",
+					Name:            "",
+					ResourceRequest: true,
+				},
+			},
+		},
+		{
+			"with http header rewrites config but missing header",
+			&authz.Config{
+				Rewrites:           &authz.SubjectAccessReviewRewrites{ByQueryParameter: &authz.QueryParameterRewriteConfig{Name: "namespace"}},
+				ResourceAttributes: &authz.ResourceAttributes{Namespace: "{{ .Value }}", APIVersion: "v1", Resource: "namespace", Subresource: "metrics"},
+			},
+			createRequest(nil, nil),
+			nil,
+		},
+		{
+			"with http header and query param rewrites config",
+			&authz.Config{
+				Rewrites: &authz.SubjectAccessReviewRewrites{
+					ByHTTPHeader:     &authz.HTTPHeaderRewriteConfig{Name: "namespace"},
+					ByQueryParameter: &authz.QueryParameterRewriteConfig{Name: "namespace"},
+				},
+				ResourceAttributes: &authz.ResourceAttributes{Namespace: "{{ .Value }}", APIVersion: "v1", Resource: "namespace", Subresource: "metrics"},
+			},
+			createRequest(map[string]string{"namespace": "tenant1"}, map[string]string{"namespace": "tenant2"}),
+			[]authorizer.Attributes{
+				authorizer.AttributesRecord{
+					User:            nil,
+					Verb:            "get",
+					Namespace:       "tenant1",
+					APIGroup:        "",
+					APIVersion:      "v1",
+					Resource:        "namespace",
+					Subresource:     "metrics",
+					Name:            "",
+					ResourceRequest: true,
+				},
+				authorizer.AttributesRecord{
+					User:            nil,
+					Verb:            "get",
+					Namespace:       "tenant2",
+					APIGroup:        "",
+					APIVersion:      "v1",
+					Resource:        "namespace",
+					Subresource:     "metrics",
+					Name:            "",
+					ResourceRequest: true,
+				},
+			},
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.desc, func(t *testing.T) {
+			t.Log(c.req.URL.Query())
+			n := krpAuthorizerAttributesGetter{authzConfig: c.authzCfg}
+			res := n.GetRequestAttributes(nil, c.req)
+			if !cmp.Equal(res, c.expected) {
+				t.Errorf("Generated authorizer attributes are not correct. Expected %v, recieved %v", c.expected, res)
+			}
+		})
+	}
+}
+
+func createRequest(queryParams, headers map[string]string) *http.Request {
+	r := httptest.NewRequest("GET", "/accounts", nil)
+	if queryParams != nil {
+		q := r.URL.Query()
+		for k, v := range queryParams {
+			q.Add(k, v)
+		}
+		r.URL.RawQuery = q.Encode()
+	}
+	if headers != nil {
+		for k, v := range headers {
+			r.Header.Set(k, v)
+		}
+	}
+	return r
 }
 
 func setupTestScenario() []testCase {