+++ test/e2e/main_test.go
@@ -49,10 +49,12 @@ func TestMain(m *testing.M) {
 
 func Test(t *testing.T) {
 	tests := map[string]kubetest.TestSuite{
-		"Basics":        testBasics(suite),
-		"TokenAudience": testTokenAudience(suite),
-		"AllowPath":     testAllowPathsRegexp(suite),
-		"IgnorePath":    testIgnorePaths(suite),
+		"Basics":             testBasics(suite),
+		"ClientCertificates": testClientCertificates(suite),
+		"TokenAudience":      testTokenAudience(suite),
+		"AllowPath":          testAllowPathsRegexp(suite),
+		"IgnorePath":         testIgnorePaths(suite),
+		"TLS":                testTLS(suite),
 	}
 
 	for name, tc := range tests {